Borders: Clear style when clearing width and color
We need to set a border style when first selecting a width or color so that there is a visible border immediately. This change will ensure that if a user manually clears either the width or color field and there is no longer a configured width or color, that the border style is also cleared.